remove get_fa function call

+++ libs/tracking/interpolation_process.cpp
@@ -22,7 +22,7 @@ bool trilinear_interpolation_with_gaussian_basis::evaluate(std::shared_ptr<track
     float ww = std::accumulate(tri_interpo.ratio,tri_interpo.ratio+8,0.0f)*0.5f;
     for (unsigned int index = 0;index < 8;++index)
     {
-        unsigned int odf_space_index = tri_interpo.dindex[index];
+        size_t odf_space_index = tri_interpo.dindex[index];
         if (!fib->get_dir(odf_space_index,ref_dir,main_dir,threshold,angle,dt_threshold))
             continue;
 		float w = tri_interpo.ratio[index];
@@ -54,7 +54,7 @@ bool trilinear_interpolation::evaluate(std::shared_ptr<tracking_data> fib,
     float total_weighting = 0.0f;
     for (unsigned int index = 0;index < 8;++index)
     {
-        int64_t odf_space_index = tri_interpo.dindex[index];
+        size_t odf_space_index = tri_interpo.dindex[index];
         if (!fib->get_dir(odf_space_index,ref_dir,main_dir,threshold,angle,dt_threshold))
             continue;
 		float w = tri_interpo.ratio[index];
@@ -77,9 +77,9 @@ bool nearest_direction::evaluate(std::shared_ptr<tracking_data> fib,
                                  float angle,
                                  float dt_threshold)
 {
-    int x = std::round(position[0]);
-    int y = std::round(position[1]);
-    int z = std::round(position[2]);
+    int x = int(std::round(position[0]));
+    int y = int(std::round(position[1]));
+    int z = int(std::round(position[2]));
     if(!fib->dim.is_valid(x,y,z))
         return false;
     if(!fib->get_dir(tipl::pixel_index<3>(x,y,z,fib->dim).index(),ref_dir,result,threshold,angle,dt_threshold))
